+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This release extends support to OneView Rest API version 1800 (OneView v5.30) fo
 - Ethernet Network
 - FC Network
 - FCOE Network
+- Interconnects
+- Logical Enclosure
+- Logical Interconnect
 - Network Set
 - Scope
 - Storage Pool